requirements.txt
affinda==1.5.0
anyio==3.6.2
asttokens==2.2.1
attrs==22.2.0
azure-core==1.26.1
azure-functions==1.12.0
azure-identity==1.12.0
backcall==0.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
comm==0.1.2
cryptography==39.0.0
debugpy==1.6.4
decorator==5.1.1
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
entrypoints==0.4
exceptiongroup==1.1.0
executing==1.2.0
fastapi==0.88.0
fastapi-microsoft-identity==0.1.6
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.1
idna==3.4
iniconfig==1.1.1
ipykernel==6.19.4
ipython==8.7.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==7.4.8
jupyter_core==5.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
msal==1.20.0
msal-extensions==1.0.0
msrest==0.7.1
multidict==6.0.4
nest-asyncio==1.5.6
oauthlib==3.2.2
orjson==3.8.3
packaging==22.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.6.2
pluggy==1.0.0
portalocker==2.6.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pur==7.0.0
pure-eval==0.2.2
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.3
Pygments==2.14.0
PyJWT==2.6.0
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.22.0
tomli==2.0.1
tornado==6.2
traitlets==5.8.0
typing_extensions==4.4.0
ujson==5.6.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
wcwidth==0.2.5
websockets==10.4
